requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
python-dotenv==0.21.0
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tornado==6.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
beautifulsoup4==4.11.1
<<<<<<< HEAD
bs4==0.0.1
=======
bs4==0.0.1
>>>>>>> upstream/main
